OpenResty deployed by Ansible with some defaults
OpenResty Ansible role created to try and replace nginx.
Copied over some configuration from the nginx role, added
some backwards and forward compatibility bits.
Still need to fabricate an init script and add the Ansible
handler for it. Hacks were done to make sure we don't
build the software over and over every time we run the role,
but that we are still able to deploy nginx configurations
willy nilly.