New AMI uses OpenResty instead of vanilla nginx
This packer builds an AMI with OpenResty as the default
web server. Is a drop in replacement for our vanilla nginx.
Configured to be compatible with existing services, and
also support use of lua modules, namely nginx-jwt for the
sake of Auth0 authentication at the web server layer.

As configured, nginx will start on boot, but will not listen
on any port, because no default server has been configured.
hostname.txt is deployed nonetheless.

Actual configurations should be deployed using ansible-deploy
which is aware of the configuration file locations.